refactor: move extensions to unified location
